CUMULUS-760 Moved API test utils to integration-tests package
