+++ packages/integration-tests/index.js
@@ -9,6 +9,8 @@ const pLimit = require('p-limit');
 const { s3, sfn } = require('@cumulus/common/aws');
 const sfnStep = require('./sfnStep');
 const { Provider, Collection, Rule } = require('@cumulus/api/models');
+
+const api = require('./api');
 const cmr = require('./cmr.js');
 
 const executionStatusNumRetries = 100;
@@ -322,6 +324,7 @@ async function buildAndExecuteWorkflow(
 }
 
 module.exports = {
+  api,
   testWorkflow,
   executeWorkflow,
   buildAndExecuteWorkflow,
